Handcalcs

Introduction

This is the documentation for Handcalcs.jl. This package supplies macros to generate $\LaTeX$ formatted strings from mathmatical formulas. This package takes inspiration from handcalcs.py which is a python package that works best in jupyter notebooks. The goal is to get the functionalities of that package and bring them to Julia. The current version of Handcalcs.jl is working for typical algebraic formulas. Future plans are to integrate the package with Unitful.jl, be able to render the algebraic expressions within a function, and many other things. This package is an extension of Latexify.jl. The @latexdefine macro is similar to the main @handcalc macro, but instead of only a symbolic rendering it also renders the numeric substitution.

Examples

A simple example for a single expression:

using Handcalcs
a = 2
b = -5
c = 2
@handcalc x = (-b + sqrt(b^2 - 4*a*c))/ (2*a)

\[x = \frac{ - b + \sqrt{b^{2} - 4 \cdot a \cdot c}}{2 \cdot a} = \frac{ + 5 + \sqrt{\left( -5 \right)^{2} - 4 \cdot 2 \cdot 2}}{2 \cdot 2} = 2.0\]

The variable x is still evaluated:

x

2.0

A simple example for multiple expressions (note the macro is now handcalcs):

b = 5 # width
h = 15 # height
@handcalcs begin
  I_x = (b*h^3)/12; "moment of inertia about x";
  I_y = (h*b^3)/12; "moment of inertia about y";
end

\[\begin{align} \\I_{x} &= \frac{b \cdot h^{3}}{12} = \frac{5 \cdot 15^{3}}{12} = 1406.25\text{ }(\text{moment of inertia about x}) \\I_{y} &= \frac{h \cdot b^{3}}{12} = \frac{15 \cdot 5^{3}}{12} = 156.25\text{ }(\text{moment of inertia about y}) \end{align}\]

The I_x and I_y variables are still evaluated:

println("The moment of inertia about the x direction is: $I_x\n
The moment of inertia about the y direction is: $I_y\n")

The moment of inertia about the x direction is: 1406.25

The moment of inertia about the y direction is: 156.25

Using Unitful with UnitfulLatexify

The package has plans to work with the packages Unitful.jl and UnitfulLatexify.jl. The only issue known is the rendering of units under exponents. Parenthesis need to wrap the numerical value and the unit for it to display correctly. See example below.

using Unitful, UnitfulLatexify
a = 2u"inch"
b = -5u"inch"
@handcalc c = sqrt(a^2 + b^2)

\[c = \sqrt{a^{2} + b^{2}} = \sqrt{2\;\mathrm{inch}^{2} -5\;\mathrm{inch}^{2}} = 5.385\;\mathrm{inch}\]

You can see that it looks as though only the unit is being squared. This should be an easy fix. However, not sure if this should be done in this package or the UnitfulLatexify package.

Future Plans

There are a number of things that I would like to implement to the package. See handcalcs.py for potential features. Here is a list of features I hope to add:

A parameters macro similar to python package

A macro that will generate $\LaTeX$ for a function that was called. The generated $\LaTeX$ would be the algebraic equations within the function

A way to break down a $\LaTeX$ equation that is too long to multiple lines

References

Handcalcs.Handcalcs

Handcalcs.@handcalc

Handcalcs.@handcalcs

Handcalcs.Handcalcs — Module

Module for better calc documentation.

Handcalcs.@handcalc — Macro

@handcalc expression

Create LaTeXString representing expression. The expression being a vaiable followed by an equals sign and an algebraic equation. Any side effects of the expression, like assignments, are evaluated as well. The RHS can be formatted or otherwise transformed by supplying a function as kwarg post.

Examples

julia> a = 2
2
julia> b = 5
5
julia> @handcalc c = a + b
L"$c = a + b = 2 + 5 = 7$"

julia> c
7

Handcalcs.@handcalcs — Macro

@handcalcs expressions

Create LaTeXString representing expressions. The expressions representing a number of expressions. A single expression being a vaiable followed by an equals sign and an algebraic equation. Any side effects of the expression, like assignments, are evaluated as well. The RHS can be formatted or otherwise transformed by supplying a function as kwarg post. Can also add comments to the end of equations. See example below.

Examples

julia> a = 2
2
julia> b = 5
5
julia> @handcalcs begin 
    c = a + b; "eq 1"
    d = a - c
end
L"$egin{align}
\\c &= a + b = 2 + 5 = 7\text{  }(\text{eq 1})
\\d &= a - c = 2 - 7 = -5
nd{align}$"

julia> c
7
julia> d
-5